requirements.txt
flask
python-gnupg
